move sudoku to `src/lib/components/Sudoku.svelte`
